requirements.txt
Django==3.1.4
flake8==3.8.4
mysqlclient==2.1.0
djangorestframework==3.12.2
django-cors-headers==3.4.0
PyJWT==1.7.1
djangorestframework-simplejwt==4.5.0
cerberus==1.3.4
itsdangerous==2.1.2
cashfree-sdk==0.1.3
